Split out validation tests to new file
