+++ packages/perseus/src/widgets/label-image/score-label-image.test.ts
@@ -1,10 +1,5 @@
 import scoreLabelImage, {scoreMarker} from "./score-label-image";
 
-import type {
-    PerseusLabelImageScoringData,
-    PerseusLabelImageUserInput,
-} from "../../validation.types";
-
 describe("scoreMarker", function () {
     it("should score correct for empty marker with no user answers", function () {
         const score = scoreMarker([], []);
@@ -59,64 +54,6 @@ describe("scoreMarker", function () {
 });
 
 describe("scoreLabelImage", function () {
-    it("should not grade non-interacted widget", function () {
-        const userInput: PerseusLabelImageUserInput = {
-            markers: [{label: "England"}, {label: "Germany"}, {label: "Italy"}],
-        } as const;
-
-        const scoringData: PerseusLabelImageScoringData = {
-            markers: [
-                {
-                    label: "England",
-                    answers: ["Mini", "Morris Minor", "Reliant Robin"],
-                },
-                {
-                    label: "Germany",
-                    answers: ["BMW", "Volkswagen", "Porsche"],
-                },
-                {
-                    label: "Italy",
-                    answers: ["Lamborghini", "Fiat", "Ferrari"],
-                },
-            ],
-        } as const;
-
-        const score = scoreLabelImage(userInput, scoringData);
-
-        expect(score).toHaveInvalidInput();
-    });
-
-    it("should not grade widget with not all markers answered", function () {
-        const userInput = {
-            markers: [
-                {label: "England", selected: ["Fiat"]},
-                {label: "Germany", selected: ["Lamborghini"]},
-                {label: "Italy"},
-            ],
-        } as const;
-
-        const scoringData = {
-            markers: [
-                {
-                    label: "England",
-                    answers: [],
-                },
-                {
-                    label: "Germany",
-                    answers: ["BMW", "Volkswagen", "Porsche"],
-                },
-                {
-                    label: "Italy",
-                    answers: ["Lamborghini", "Fiat", "Ferrari"],
-                },
-            ],
-        } as const;
-
-        const score = scoreLabelImage(userInput, scoringData);
-
-        expect(score).toHaveInvalidInput();
-    });
-
     it("should grade as incorrect for widget with no answers for markers", function () {
         const userInput = {
             markers: [
